// src/types/Element.interface.ts
interface Element {
    atomicMass: string;
    atomicNumber: string;
    atomicRadius: string;
    boilingPoint: string;
    bondingType: string;
    cpkHexColor: string;
    density: string;
    electronAffinity: string;
    electronegativity: string;
    electronicConfiguration: string;
    groupBlock: string;
    ionRadius: string;
    ionizationEnergy: string;
    meltingPoint: string;
    name: string;
    oxidationStates: string;
    standardState: string;
    symbol: string;
    vanDelWaalsRadius: string;
    yearDiscovered: string;
}

export type { Element };